+++ lib/truffle/truffle/cext.rb
@@ -683,7 +683,7 @@ module Truffle::CExt
   end
 
   def rb_set_safe_level_force(level)
-    Truffle::ThreadOperations.set_thread_local( :$SAFE, level)
+    $SAFE = level
   end
 
   def rb_thread_alone