+++ src/main/ruby/core/type.rb
@@ -646,21 +646,6 @@ module Truffle
       end
     end
 
-    def self.check_safe_level(level)
-      level = rb_num2int(level)
-      current_level = Thread.current.safe_level
-      if level < current_level
-        raise SecurityError, "tried to downgrade safe level from #{current_level} to #{level}"
-      end
-      if level > 1 # SAFE_LEVEL_MAX
-        raise ArgumentError, '$SAFE=2 to 4 are obsolete'
-      end
-      if level != 0
-        Truffle::Debug.log_warning '$SAFE level checks not implemented'
-      end
-      level
-    end
-
     # Needs to be in core for assigning $!
     def self.set_last_exception(error)
       if !error.nil? && !error.is_a?(Exception)