+++ test/mri/excludes/TestGemSpecification.rb
@@ -1,1 +1,2 @@
 exclude :test_default_spec_stub_is_marked_default, "needs investigation"
+exclude :test_self_load_tainted, "safe levels not supported"