+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Bug fixes:
 * Support `Signal.trap('PIPE', 'SYSTEM_DEFAULT')`, used by the gem `rouge`
   (#1411).
 * Fix arity checks and handling of arity `-2` for `rb_define_method()`.
+* Setting `$SAFE` to a negative value now raises a `SecurityError`.
 
 Performance:
 
@@ -46,6 +47,9 @@ Changes:
 * The `-Xembedded` option can now be set set on the launcher command line.
 * The `-Xplatform.native=false` option can now load the core library, by
   enabling `-Xpolyglot.stdio`.
+* `$SAFE` and `Thread#safe_level` now cannot be set to `1` - raising an error
+  rather than warning as before. `-Xsafe` allows it to be set, but there are
+  still no checks.
 
 Changes: